Move export and missingfilesdialog to io.py
